refactored for testing
